Extract ActiveRecord::ConnectionAdapters::OracleEnhanced::Quoting
